requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
psycopg2-binary>=2.9.9
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.4
gensim>=4.3.2
scikit-learn==1.6.1
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
xlrd==2.0.1
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Migrate==4.0.5
plotly>=5.17.0
seaborn>=0.13.0
matplotlib>=3.8.0

celery==5.3.1
redis==4.6.0
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Gunicorn==21.2.0
python-dotenv==1.0.0
schedule==1.2.0
python-magic==0.4.27
Sastrawi>=1.0.1
bleach==6.1.0
chardet==5.2.0
Flask-CORS==4.0.0
watchdog>=3.0.0
